% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircumOLS.R
\name{CircumOLS}
\alias{CircumOLS}
\title{CirucmOLS}
\usage{
CircumOLS(
  rawdt,
  m = 1,
  mcsc = "unconstrained",
  type = "ordinal",
  simulation_based = T,
  icorrection = 0,
  ncore = 8
)
}
\arguments{
\item{rawdt}{requires data set.}

\item{m}{indicates the number of beta coefficients}

\item{mcsc}{indicates the minimum common score correlation}

\item{type}{indicates the type of data}

\item{simulation_based}{determines the type of ACM}

\item{icorrection}{indicates the type of correction for polychoric correlation}

\item{ncore}{indicates the number of cores for parallel computing.}
}
\value{
ACM LLALALAL
}
\description{
CirucmOLS
}
