% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircumOLS.R
\name{CircumOLS}
\alias{CircumOLS}
\title{Explanation}
\usage{
CircumOLS(
  rawdt,
  m = 1,
  mcsc = "unconstrained",
  type = "ordinal",
  simulation_based = T,
  icorrection = 0,
  ncore = 8
)
}
\value{

}
\description{
Explanation
}
